+++ doc/changelog.rst
@@ -9,7 +9,7 @@ All notable changes to this project will be documented in this file. The format
 is based on `Keep a Changelog <https://keepachangelog.com/en/1.1.0>`_, and this
 project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
-Contributors to each release are listed in alphabetical order by first name. 
+Contributors to each release are listed in alphabetical order by first name.
 List entries are sorted in descending chronological order.
 
 Unreleased
@@ -21,8 +21,11 @@ Contributors
 
 Added
 -----
+- Projection/pattern center calibration via the moving screen technique in a
+  kikuchipy.detectors.calibration module.
+  (`#322 <https://github.com/pyxem/kikuchipy/pull/322>`_)
 - Three single crystal Si EBSD patterns, from the same sample position but with
-  varying sample-screen distances, to the data module (via external repo).
+  varying detector distances, to the data module (via external repo).
   (`#320 <https://github.com/pyxem/kikuchipy/pull/320>`_)
 - Reading of NORDIF calibration patterns specified in a setting file into an
   EBSD signal. (`#317 <https://github.com/pyxem/kikuchipy/pull/317>`_)